refactor: construct a pattern in a single pass
The previous implementation iterated multiple times over the guess and
the solution in order to handle every edge case, while mutating the map
of characters by position.
By counting the characters in the solution instead, and decrementing the
count whenever a character is used, we can construct the pattern in a
single pass.